Refactor functionality out of character cog This allows for easier helpdoc implementation and unit testing
