+++ cogs/character.py
@@ -9,64 +9,8 @@ class Character(commands.Cog):
         self.client = client
         self.dataPath = os.getenv("data")
         self.font = urlopen('https://github.com/google/fonts/raw/main/ufl/ubuntu/Ubuntu-Medium.ttf')
-        self.loadData()
-
-    def loadData(self):
-        try:
-            data = urlopen(f'{self.dataPath}/characters.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'characters.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.chars = dict((k.lower(), v) for k, v in json.load(data).items())
-
-        try:
-            data = urlopen(f'{self.dataPath}/emoji.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'emoji.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.emojis = json.load(data)
-
-        try:
-            data = urlopen(f'{self.dataPath}/ougi.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'ougi.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.ougis = dict((k.lower(), v) for k, v in json.load(data).items())
-
-        try:
-            data = urlopen(f'{self.dataPath}/skill.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'skill.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.skills = dict((k.lower(), v) for k, v in json.load(data).items())
-
-        try:
-            data = urlopen(f'{self.dataPath}/supportskill.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'supportskill.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.supportSkills = dict((k.lower(), v) for k, v in json.load(data).items())
-
-        try:
-            data = urlopen(f'{self.dataPath}/emp.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'emp.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.emps = dict((k.lower(), v) for k, v in json.load(data).items())
-
-        try:
-            data = urlopen(f'{self.dataPath}/empdata.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'empdata.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.empData = json.load(data)
-
-        try:
-            data = urlopen(f'{self.dataPath}/empdomaindata.json')
-        except Exception:
-            with open(os.path.join(self.dataPath, 'empdomaindata.json'), 'r', encoding='utf-8') as dataFile:
-                data = io.StringIO(dataFile.read())
-        self.empDomainData = json.load(data)
+        self.icon_url = 'https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png'
+        self.helper = client.get_cog('CharHelper')
 
     @commands.Cog.listener()
     async def on_ready(self):
@@ -87,7 +31,7 @@ class Character(commands.Cog):
 
     @char.command()
     async def reload(self, ctx):
-        self.loadData()
+        self.helper.loadData()
         await ctx.send('Data reloaded')
 
     @char.command()
@@ -97,54 +41,25 @@ class Character(commands.Cog):
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
-        await self.sendUncap(ctx, name, version, uncap)
-        charVersion = self.chars[name][version]
-
-        name = name.title()
-        version = version.title()
-
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
+        msg = self.getUncapMessage(name, version, uncap)
+        mainEmbed = self.helper.getInfo(name, version)
         embedList = []
-
-        #Main embed
-        title = f'{self.emojis["Rarity"][charVersion["rarity"].upper()]} '
-        for series in charVersion['series']:
-            title += f'{self.emojis["Series"][series]} '
-        title +=f'**{charVersion["name"]}**'
-        description = f'**JP**: {charVersion["jpname"]}\n'
-        description += f'**VA**: {", ".join(charVersion["va"])}'
-
-        information = f'\n{self.emojis["Element"][charVersion["element"]]} **{charVersion["element"]}** {self.emojis["Type"][charVersion["type"]]} **{charVersion["type"]}**\n'
-        for specialty in charVersion['specialties']:
-            information += f'{self.emojis["Specialty"][specialty]} **{specialty}**  '
-        information += '\n'
-        for gender in charVersion['gender']:
-            information += f'{self.emojis["Gender"][gender]} **{gender}**  '
-        information += '\n'
-        for race in charVersion['race']:
-            information += f'{self.emojis["Race"][race]} **{race}**  '
-        information += f'\n{self.emojis["Stat"]["Hp"]} *{charVersion["hp"]}* {self.emojis["Stat"]["Atk"]} *{charVersion["atk"]}*'
-
-        profile = f'{charVersion["profile"]}'
-
-        mainEmbed = discord.Embed()
-        mainEmbed.title = title
-        mainEmbed.description = description
-        mainEmbed.add_field(name='Basic Info.', value=information, inline=False)
-        mainEmbed.add_field(name='Profile', value=profile, inline=False)
-        mainEmbed.set_thumbnail(url=charVersion['thumbnail'])
-        mainEmbed.set_image(url=charVersion['image'])
         embedList.append(mainEmbed)
-
-        embedList.append(await self.ougi(ctx, name, version, uncap, noShow=True))
-        embedList.extend(await self.skill(ctx, name, version, uncap, noShow=True))
-        embedList.extend(await self.support(ctx, name, version, uncap, noShow=True))
+        embedList.append(self.helper.getOugi(name, version))
+        embedList.extend(self.helper.getSkill(name, version))
+        embedList.extend(self.helper.getSupport(name, version))
         #TODO: Look into ImgurAPI for this
         # embedList.append(await self.emp(ctx, name, version, uncap, noShow=True))
 
         for embed in embedList:
             index = embedList.index(embed)
-            embed.set_footer(text=f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki', icon_url='https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png')
+            footerText = f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki'
+            if msg:
+                footerText += f'\n{msg}'
+            embed.set_footer(text=footerText, icon_url=self.icon_url)
         paginator = DiscordUtils.Pagination.CustomEmbedPaginator(ctx, timeout=60, remove_reactions=True, auto_footer=False)
         paginator.add_reaction('⏮️', "first")
         paginator.add_reaction('⏪', "back")
@@ -155,112 +70,53 @@ class Character(commands.Cog):
         await paginator.run(embedList)
 
     @char.command()
-    async def ougi(self, ctx, name :str, version=None, uncap='6', noShow=False):
+    async def ougi(self, ctx, name :str, version=None, uncap='6'):
         name, version, uncap, noVersion = self.getCharVersion(ctx, name, version, uncap)
         if not name:
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
-        if not noShow:
-            await self.sendUncap(ctx, name, version, uncap)
-        charVersion = self.ougis[name][version]
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
+        msg = self.getUncapMessage(name, version, uncap)
 
-        ougiEmbed = discord.Embed()
-        for ougiList in charVersion:
-            value=''
-            for ougiText in ougiList["text"]:
-                value+= ougiText +"\n"
-            ougiEmbed.add_field(name=f'**{ougiList["name"]}**:', value=value, inline=False)
+        ougiEmbed = self.helper.getOugi(name, version)
 
-            if "data" in ougiList:
-                data = ougiList["data"]
-                for details in data:
-                    ougiEmbed.add_field(name=details['title'], value="\n".join(details['text']), inline=details['inLine'])
-
-            if ougiList["duration"]:
-                duration = ''
-                for ougiDuration in ougiList["duration"]:
-                    for ougiDurationText in ougiList["duration"][ougiDuration]:
-                            duration+= f'{ougiDurationText} and '
-                    duration = duration[:-4]
-                    duration += f': {ougiDuration}.\n'
-                    duration = duration.replace('^s', ' seconds')
-                    duration = duration.replace('^i', '')
-                    if ougiDuration == '1^t':
-                        duration = duration.replace('^t', ' turn')
-                    else:
-                        duration = duration.replace('^t', ' turns')
-                ougiEmbed.add_field(name='Durations', value=f'{duration}\n', inline=False)
-
-        ougiEmbed.title="Charge Attack"
-        ougiEmbed.set_thumbnail(url='https://cdn.discordapp.com/attachments/828230361321963530/830390392565923900/download.png')
-        ougiEmbed.set_image(url=self.chars[name][version]['image'])
-
-        if noShow:
-            return ougiEmbed
-        else:
-            ougiEmbed.set_footer(text=f'Data obtained from GBF Wiki', icon_url='https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png')
-            await ctx.send(embed=ougiEmbed)
+        footerText = f'Data obtained from GBF Wiki'
+        if msg:
+            footerText += f'\n{msg}'
+        ougiEmbed.set_footer(text=footerText, icon_url=self.icon_url)
+        await ctx.send(embed=ougiEmbed)
 
     @char.command()
-    async def skill(self, ctx, name, version=None, uncap='6', noShow=False):
+    async def skill(self, ctx, name, version=None, uncap='6'):
         name, version, uncap, noVersion = self.getCharVersion(ctx, name, version, uncap)
         if not name:
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
-        if not noShow:
-            await self.sendUncap(ctx, name, version, uncap)
-        charVersion = self.skills[name][version]
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
+        msg = self.getUncapMessage(name, version, uncap)
 
-        embedList = []
-        for skillList in charVersion:
-            if len(skillList) == 0: continue
+        embedList = self.helper.getSkill(name, version)
 
-            for skill in skillList:
-                skillEmbed = discord.Embed(title=f'{skill["name"]}')
-                skillEmbed.add_field(name=f'Cooldown: {skill["cooldown"]}', value='\n'.join(skill['text']), inline=False)
-                skillEmbed.set_thumbnail(url=f'{skill["icon"][-1]}')
-                skillEmbed.set_image(url=self.chars[name][version]['image'])
+        for embed in embedList:
+            index = embedList.index(embed)
+            footerText = f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki'
+            if msg:
+                footerText += f'\n{msg}'
+            embed.set_footer(text=footerText, icon_url=self.icon_url)
+        paginator = DiscordUtils.Pagination.CustomEmbedPaginator(ctx, timeout=60, remove_reactions=True, auto_footer=False)
+        paginator.add_reaction('⏮️', "first")
+        paginator.add_reaction('⏪', "back")
+        paginator.add_reaction('🔐', "lock")
+        paginator.add_reaction('⏩', "next")
+        paginator.add_reaction('⏭️', "last")
 
-                if "data" in skill:
-                    data = skill["data"]
-                    for details in data:
-                        skillEmbed.add_field(name=details['title'], value="\n".join(details['text']), inline=details['inLine'])
-
-                if skill['duration']:
-                    duration = ''
-                    for skillDuration in skill['duration']:
-                        for skillDurationText in skill["duration"][skillDuration]:
-                                duration+= f'{skillDurationText} and '
-                        duration = duration[:-4]
-                        duration += f': {skillDuration}.\n'
-                        duration = duration.replace('^s', ' seconds')
-                        duration = duration.replace('^i', '')
-                        if skillDuration == '1^t':
-                            duration = duration.replace('^t', ' turn')
-                        else:
-                            duration = duration.replace('^t', ' turns')
-                    skillEmbed.add_field(name='Durations', value=f'{duration}\n', inline=False)
-
-            embedList.append(skillEmbed)
-
-        if noShow:
-            return embedList
-        else:
-            for embed in embedList:
-                index = embedList.index(embed)
-                embed.set_footer(text=f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki', icon_url='https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png')
-            paginator = DiscordUtils.Pagination.CustomEmbedPaginator(ctx, timeout=60, remove_reactions=True, auto_footer=False)
-            paginator.add_reaction('⏮️', "first")
-            paginator.add_reaction('⏪', "back")
-            paginator.add_reaction('🔐', "lock")
-            paginator.add_reaction('⏩', "next")
-            paginator.add_reaction('⏭️', "last")
-
-            await paginator.run(embedList)
+        await paginator.run(embedList)
 
     @char.command()
     async def support(self, ctx, name :str, version=None, uncap='6', noShow=False):
@@ -269,133 +125,48 @@ class Character(commands.Cog):
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
-        if not noShow:
-            await self.sendUncap(ctx, name, version, uncap)
-        charVersion = self.supportSkills[name][version]
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
+        msg = self.getUncapMessage(name, version, uncap)
 
-        embedList = []
-        for supportList in charVersion:
-            supportEmbed = discord.Embed()
-            supportEmbed.title = f'{supportList["name"]}:'
-            supportEmbed.description = "\n".join(supportList['text'])
-            supportEmbed.set_thumbnail(url=f'{supportList["thumbnail"]}')
-            supportEmbed.set_image(url=self.chars[name][version]['image'])
+        embedList = self.helper.getSupport(name, version)
 
-            if "data" in supportList:
-                data = supportList["data"]
-                for details in data:
-                    supportEmbed.add_field(name=details['title'], value="\n".join(details['text']), inline=details['inLine'])
+        for embed in embedList:
+            index = embedList.index(embed)
+            footerText = f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki'
+            if msg:
+                footerText += f'\n{msg}'
+            embed.set_footer(text=footerText, icon_url=self.icon_url)
+        paginator = DiscordUtils.Pagination.CustomEmbedPaginator(ctx, timeout=60, remove_reactions=True, auto_footer=False)
+        paginator.add_reaction('⏮️', "first")
+        paginator.add_reaction('⏪', "back")
+        paginator.add_reaction('🔐', "lock")
+        paginator.add_reaction('⏩', "next")
+        paginator.add_reaction('⏭️', "last")
 
-            if supportList["duration"]:
-                duration = ''
-                for supportDuration in supportList["duration"]:
-                    for supportDurationText in supportList["duration"][supportDuration]:
-                            duration+= f'{supportDurationText} and '
-                    duration = duration[:-4]
-                    duration += f': {supportDuration}.\n'
-                    duration = duration.replace('^s', ' seconds')
-                    duration = duration.replace('^i', '')
-                    if supportDuration == '1^t':
-                        duration = duration.replace('^t', ' turn')
-                    else:
-                        duration = duration.replace('^t', ' turns')
-                supportEmbed.add_field(name='Durations', value=f'{duration}\n', inline=False)
-            embedList.append(supportEmbed)
-
-        if noShow:
-            return embedList
-        else:
-            for embed in embedList:
-                index = embedList.index(embed)
-                embed.set_footer(text=f'({index+1}/{len(embedList)})\nData obtained from GBF Wiki', icon_url='https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png')
-            paginator = DiscordUtils.Pagination.CustomEmbedPaginator(ctx, timeout=60, remove_reactions=True, auto_footer=False)
-            paginator.add_reaction('⏮️', "first")
-            paginator.add_reaction('⏪', "back")
-            paginator.add_reaction('🔐', "lock")
-            paginator.add_reaction('⏩', "next")
-            paginator.add_reaction('⏭️', "last")
-
-            await paginator.run(embedList)
+        await paginator.run(embedList)
 
     @char.command()
     async def emp(self, ctx, name : str, version=None, uncap='6', noShow=False):
-        from PIL import Image, ImageDraw, ImageFont
-        import urllib.request, os
+        
 
         name, version, uncap, noVersion = self.getCharVersion(ctx, name, version, uncap)
         if not name:
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
-        if not noShow:
-            await self.sendUncap(ctx, name, version, uncap)
-        charVersion = self.emps[name][version]
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
+        msg = self.getUncapMessage(name, version, uncap)
+        
+        embed, file = self.helper.getEmp(name, version)
+        footerText = f'Data obtained from GBF Wiki'
+        if msg:
+            footerText += f'\n{msg}'
+        embed.set_footer(text=footerText, icon_url=self.icon_url)
 
-        try:
-            file = discord.File(f'cache/emp/char/{name}/{version}.png', filename='emp.png')
-        except Exception:
-            image = Image.new("RGBA", (104*5, 104*len(charVersion)))
-            i = 0
-            for row in charVersion:
-                j = 0
-                rowImage = Image.new("RGBA", (104*5, 104))
-                for emp in row:
-                    try:
-                        id = self.empData[emp]
-                        try:
-                            cellImage = Image.open(f'cache/emp/image/normal/{id}.png')
-                            rowImage.paste(cellImage, (104*j, 0))
-                        except Exception:
-                            os.makedirs('cache/emp/image/normal', exist_ok=True)
-                            url = f'http://game-a.granbluefantasy.jp/assets_en/img/sp/zenith/assets/ability/{id}.png'
-                            cellImage = Image.open(urllib.request.urlopen(url))
-                            cellImage.save(f'cache/emp/image/normal/{id}.png')
-                            rowImage.paste(cellImage, (104*j, 0))
-                    except:
-                        id = self.empDomainData[emp]
-                        if id == 0:
-                            try:
-                                cellImage = Image.open(f'cache/emp/image/domain/0.png')
-                                rowImage.paste(cellImage, (104*j, 0))
-                            except Exception:
-                                os.makedirs('cache/emp/image/domain', exist_ok=True)
-                                cellImage= Image.new("RGBA", (520, 104))
-                                domainText = ImageDraw.Draw(cellImage)
-                                font = ImageFont.truetype(self.font, 50)
-                                domainText.text((258,52), "Domain of the Evoker", anchor= "mm", font=font)
-                                cellImage.save(f'cache/emp/image/domain/{id}.png')
-                                rowImage.paste(cellImage, (104*j, 0))
-
-                        else:
-                            try:
-                                cellImage = Image.open(f'cache/emp/image/domain/{id}.png')
-                                rowImage.paste(cellImage, (104*j, 0))
-                            except Exception:
-                                os.makedirs('cache/emp/image/normal', exist_ok=True)
-                                url = f'http://game-a.granbluefantasy.jp/assets_en/img/sp/domain_evoker/assets/domain_icon/{id}/i_released.png'
-                                cellImage = Image.open(urllib.request.urlopen(url))
-                                cellImage= cellImage.crop((10, 10, 114, 114))
-                                cellImage.save(f'cache/emp/image/domain/{id}.png')
-                                rowImage.paste(cellImage, (104*j, 0))
-
-                    j += 1
-                image.paste(rowImage, (0, 104*i))
-                i += 1
-
-            os.makedirs(f'cache/emp/char/{name}', exist_ok=True)
-            image.save(f'cache/emp/char/{name}/{version}.png', format="PNG")
-            file = discord.File(f'cache/emp/char/{name}/{version}.png', filename="emp.png")
-
-        embed = discord.Embed()
-        embed.title = '__Extended Mastery Perks__'
-        embed.set_thumbnail(url=self.chars[name][version]['thumbnail'])
-        embed.set_image(url="attachment://emp.png")
-        embed.set_footer(text=f'Data obtained from GBF Wiki', icon_url='https://cdn.discordapp.com/attachments/828230402875457546/839701583515222026/321247751830634496.png')
-
-        if noShow:
-            return embed
         await ctx.send(file=file, embed=embed)
 
     @char.command(hidden=True)
@@ -406,12 +177,14 @@ class Character(commands.Cog):
 
     @char.command()
     async def art(self, ctx, name, version=None):
-        name, version, uncap, noVersion = self.getCharVersion(ctx, name, version, None)
+        name, version, _, noVersion = self.getCharVersion(ctx, name, version, None)
         if not name:
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
         charVersion = self.chars[name][version]
         charId = charVersion['id']
         maxVersion = 2
@@ -451,17 +224,15 @@ class Character(commands.Cog):
         import sqlite3
         from contextlib import closing
 
-        name, versionTemp, uncap, noVersion = self.getCharVersion(ctx, name, version)
+        name, versionTemp, _, _ = self.getCharVersion(ctx, name, version)
         if not name:
             await ctx.send('Character not found!')
             return
-        if noVersion:
-            await self.sendDefault(ctx, name)
         alias = alias.lower()
         connection = sqlite3.connect('db/kolulu.db')
         with closing(connection) as db:
             cursor = db.cursor()
-            char = self.chars[name]
+            char = self.helper.chars[name]
             if versionTemp == 'BASE' and not version:
                 try:
                     statement = 'INSERT INTO alt_names VALUES(?, ?, ?)'
@@ -513,12 +284,14 @@ class Character(commands.Cog):
         import sqlite3
         from contextlib import closing
 
-        name, version, uncap, noVersion = self.getCharVersion(ctx, name, version, None)
+        name, version, _, noVersion = self.getCharVersion(ctx, name, version, None)
         if not name:
             await ctx.send('Character not found!')
             return
         if noVersion:
-            await self.sendDefault(ctx, name)
+            msg = self.sendDefault(ctx, name)
+            if msg:
+                await ctx.send(msg)
 
         connection = sqlite3.connect('db/kolulu.db')
         with closing(connection) as db:
@@ -566,7 +339,7 @@ class Character(commands.Cog):
         elif uncap.upper() == 'ULB':
             uncap = '6'
 
-        if not name in self.chars:
+        if not name in self.helper.chars:
             import sqlite3
             from contextlib import closing
             connection = sqlite3.connect('db/kolulu.db')
@@ -587,7 +360,7 @@ class Character(commands.Cog):
                     else:
                         return None, version, uncap, noVersion
 
-        char = self.chars[name]
+        char = self.helper.chars[name]
         if not version or version not in char:
             version = 'BASE'
             noVersion = True
@@ -599,14 +372,17 @@ class Character(commands.Cog):
                 version += '_5'
         return name, version, uncap, noVersion
 
-    async def sendDefault(self, ctx, name):
-        versions = ', '.join((ver.title() for ver in self.chars[name] if '_' not in ver))
-        charName = self.chars[name]['BASE']['name']
+    def sendDefault(self, ctx, name):
+        versions = [version.title() for version in self.helper.chars[name] if '_' not in version]
+        versions = [version for version in versions if not version.startswith('Base')]
+        if len(versions) == 0: return ""
+        versionText = ', '.join(versions)
+        charName = self.helper.chars[name]['BASE']['name']
         charName = charName.rsplit(' ')[0] if charName.endswith(')') or charName.endswith('\u2605') else charName
-        await ctx.send(f'Version not specified or not found for character **{charName}**, using the default version. \nValid versions for character **{charName}: {versions}**')
+        return f'Using the default version for character **{charName}**.\nOther valid versions: **{versionText}**'
 
-    async def sendUncap(self, ctx, name, version, uncap):
-        charVersion = self.chars[name][version]
+    def getUncapMessage(self, name, version, uncap):
+        charVersion = self.helper.chars[name][version]
         charName = charVersion['name']
         charName = charName.rsplit(' ')[0] if charName.endswith(')') or charName.endswith('\u2605') else charName
         maxUncap = int(charVersion["max_evo"])
@@ -615,13 +391,13 @@ class Character(commands.Cog):
         if currentUncap > maxUncap:
             currentUncap = maxUncap
 
-        if maxUncap < 5: return
+        if maxUncap < 5: return ''
 
         if maxUncap == currentUncap:
-            msg = f'Showing the highest uncap for character **{charName}**. For lower uncap add 5 or 4 to the end of the command.'
+            msg = f'Showing the highest uncap for character {charName}.\nFor lower uncap add 5 or 4 to the end of the command.'
         else:
-            msg = f'Showing the {currentUncap}* uncap for character **{charName}**.'
-        await ctx.send(msg)
+            msg = f'Showing the {currentUncap}* uncap for character {charName}.'
+        return msg
 
 def setup(client):
     client.add_cog(Character(client))
